refactor: move file-utils functions to utils
